requirements.txt
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
werkzeug==2.0.3
flask==2.0.3
fastapi==0.78.0
uvicorn==0.17.6
requests==2.27.1
python-dotenv==0.20.0
pydantic==1.9.1
jinja2==3.1.2
langchain==0.1.4
langchain-openai==0.0.5
